Authentication without reviewer

+++ routes/api/admins.js
@@ -11,6 +11,18 @@ var jwt = require("jsonwebtoken");
 
 router.get("/", async (req, res) => {
   try {
+    var stat = 0;
+    var token = req.headers["x-access-token"];
+    if (!token)
+      return res
+        .status(401)
+        .send({ auth: false, message: "Please login first." });
+    jwt.verify(token, config.secret, async function(err, decoded) {
+      if (err)
+        return res
+          .status(500)
+          .send({ auth: false, message: "Failed to authenticate token." });
+    });
     const info = [];
     const arrayOfAdmins = await Admin.find();
     for (var i = 0; i < arrayOfAdmins.length; i++) {
@@ -28,13 +40,25 @@ router.get("/", async (req, res) => {
     }
     res.send(info);
   } catch (error) {
-    console.log(error);
+    res.status(404).send({ msg: "Admin doesn't exist" });
   }
 });
 
 router.get("/:id", async (req, res) => {
   try {
     const adminId = req.params.id;
+    var stat = 0;
+    var token = req.headers["x-access-token"];
+    if (!token)
+      return res
+        .status(401)
+        .send({ auth: false, message: "Please login first." });
+    jwt.verify(token, config.secret, async function(err, decoded) {
+      if (err)
+        return res
+          .status(500)
+          .send({ auth: false, message: "Failed to authenticate token." });
+    });
     const admin = await Admin.findById(adminId);
     curr = {
       name: admin.name,
@@ -47,31 +71,50 @@ router.get("/:id", async (req, res) => {
     };
     res.send(curr);
   } catch (error) {
-    console.log(error);
+    res.status(404).send({ msg: "Admin doesn't exist" });
   }
 });
 
-router.put('/:id', async (req, res) => {
-    try{
-        const isValidated = validator.updateValidation(req.body);
-        if(isValidated.error) 
-        {
-            return  res.status(400).send({error: isValidated.error.details[0].message});
-        }
-        const id = req.params.id;
-        const updatedAdmin = await Admin.FindByIdAndUpdate(id,req.body);
-        // res.send(admin);
-        res.json({msg: "Information updated successfully"});
+router.put("/:id", async (req, res) => {
+  try {
+    const id = req.params.id;
+    var stat = 0;
+    var token = req.headers["x-access-token"];
+    if (!token)
+      return res
+        .status(401)
+        .send({ auth: false, message: "Please login first" });
+    jwt.verify(token, config.secret, async function(err, decoded) {
+      if (err)
+        return res
+          .status(500)
+          .send({ auth: false, message: "Failed to authenticate token." });
+      stat = decoded.id;
+    });
+    if (id === stat) {
+      const isValidated = validator.updateValidation(req.body);
+      if (isValidated.error) {
+        return res
+          .status(400)
+          .send({ error: isValidated.error.details[0].message });
+      }
+      const updatedAdmin = await Admin.findByIdAndUpdate(id, req.body);
+      // res.send(admin);
+      res.json({ msg: "Information updated successfully" });
+    } else {
+      res.status(401).send({
+        auth: false,
+        message: "You don't have the authorization for this."
+      });
     }
-    catch(error)
-    {
-        console.log(error);
-    }
+  } catch (error) {
+    res.status(404).send({ msg: "Admin doesn't exist" });
+  }
 });
 
 //Register admin by another admin
 router.post("/", async (req, res) => {
-    var stat=0;
+  var stat = 0;
   try {
     var token = req.headers["x-access-token"];
     if (!token)
@@ -79,13 +122,12 @@ router.post("/", async (req, res) => {
         .status(401)
         .send({ auth: false, message: "Please login first." });
     jwt.verify(token, config.secret, async function(err, decoded) {
-      if (err)
-        stat=decoded.id;
-        return res
-          .status(500)
-          .send({ auth: false, message: "Failed to authenticate token." });
+      if (err) stat = decoded.id;
+      return res
+        .status(500)
+        .send({ auth: false, message: "Failed to authenticate token." });
     });
-    const admin2 =  await Admin.findById(stat);
+    const admin2 = await Admin.findById(stat);
     if (!admin2) {
       return res.status(400).json({ error: "You are not an admin" });
     }
@@ -119,11 +161,30 @@ router.post("/", async (req, res) => {
 
 router.delete("/:id", async (req, res) => {
   try {
+    var stat = 0;
+    var token = req.headers["x-access-token"];
+    if (!token)
+      return res
+        .status(401)
+        .send({ auth: false, message: "Please login first" });
+    jwt.verify(token, config.secret, async function(err, decoded) {
+      if (err)
+        return res
+          .status(500)
+          .send({ auth: false, message: "Failed to authenticate token." });
+      stat = decoded.id;
+    });
     const id = req.params.id;
-    const admin = await Admin.findByIdAndRemove(id);
-    res.json({ msg: "Admin deleted successfully" });
+    if (stat === id) {
+      const admin = await Admin.findByIdAndRemove(id);
+      res.json({ msg: "Admin deleted successfully" });
+    } else {
+      res
+        .status(401)
+        .send({ auth: false, message: "You don't have the authorization" });
+    }
   } catch (error) {
-    console.log(error);
+    res.status(404).send({ msg: "Admin doesn't exist" });
   }
 });
